+++ _install/library/modules/democontent/manifest.php
@@ -2,7 +2,6 @@
 	'label' 			=> 'Demo Content',
 	'requiresModules' => [ ],
 	'files' => [
-		'fonts',
 		'images',
 		'favicon',
 	]